+++ server/src/main/java/com/flightcoordinator/server/service/CrewService.java
@@ -82,7 +82,7 @@ public class CrewService {
     crewRepository.delete(existingCrewMember);
   }
 
-  private Boolean isPhoneNumberValid(Integer phoneNumber) {
+  private Boolean isPhoneNumberValid(Long phoneNumber) {
     return phoneNumber.toString().length() == 11;
   }
